+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 
 ### Features
 
+
+- compute: add VPC support (#880)
 ### Bug fixes
 
 ### Improvements